Created basic website generation using templates
Running generate.py will compile all .html files
in the base folder with the provided templates
and write the output HTML files to the output
folder. Also the js, css, img, and file folders
are copied to the output folder.

The website using the Foundation framework. All
JS libraries and included in the build, but pages
will only load linked files.

The README.md now provides a broad overview of
the project.